feat: class 컴포넌트 - 썸네일 추가
